ASCII only
I searched through a bunch of repos in the wild for non-ASCII characters
and most of them had none. SDL2 did have characters like 牛, 肉, 西, 蘭
for testing purposes. I imagine that I would use escape sequences for
that stuff in my own code.

I also found © and ® in some license headers but that's about it really.